Check the inside-string syntax flag rather comment/string-start flag
The 8th slot of `syntax-ppss` is "character address of start of comment
or string; nil if not in one" according to the manual, so normal comments
were not being highlighted.

`outline-minor-faces--syntactic-matcher` has been tweaked slightly to check
the 3rd slot of the result which is "non-nil if inside a string." It also
now checks if the start of the match is in a string, since even the default
outline-minor-faces-regexp for Elisp matches to the end of the line, so
point will be at the end of the line where single-line strings have already
ended.

This code matches all wanted headings, does not match in single-line strings,
does not match in multiline strings, and does not match in docstrings.