Use derived regexp in lisp modes
The default value of `outline-regexp' in various lisp modes is either
  ";;;\\(;* [^ \t\n]\\|###autoload\\)\\|("
of
  ";;;\\(;* [^ \t\n]\\)\\|("

In addition to outline headings the regexp is also supposed to match
top-level s-expressions (or rather lines beginning with "(") and in
some case also autoload cookies.

That means that we cannot use the regexp as is because, unlike the
`outline' package, we only want to match outline headings.  We
therefore try to use a regexp that only derives from `outline-regexp'.

Some maintainers don't like the convention of using three or more
semicolons for outline headings and use asterisks instead, in which
case a regexp like e.g. ";;\\(\\*+ [^ \t\n]\\|###autoload\\)\\|("
should be used.

For lisp modes we now try to detect regexps that have been modified
like so.  If that fails then we continue to use a regexp that derives
from the default value instead of one that derives from the actual
local value.  In some cases that might be about the right thing to do
but we can't really know.  The worst that can happen is that we fail
to highlight the outline headings (which previously was more common).